Extract BaseCoverArtDelegate from CoverArtDelegate
